third_party/WebKit/LayoutTests/media/video-source-type-params.html

<!DOCTYPE html>
<title>Test "source" element "type" attribute with "params".</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video>
    <source src=content/bogus.mpeg type="video/blahblah">
    <source src=content/test-vp9-opus.webm type="video/webm; codecs=&quot;vp9,opus&quot;">
    <source src=content/error2.mpeg type=video/mpeg>
</video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");

    video.onloadstart = t.step_func_done(function() {
        var url = video.currentSrc;
        assert_equals(url.substr(url.lastIndexOf("/media/") + 7), "content/test-vp9-opus.webm");
    });
});
</script>
